better HTTP CODE

+++ src/Service/UserService.php
@@ -11,14 +11,17 @@ class UserService
     public function __construct(
         private UserRepository $userRepository
     ){}
-    public function delete(User $user,Client $client)
+    public function delete(User $user,Client $client):User|false
     {
-
+        if(!$client->getUsers()->contains($user)){
+            return false;
+        }
         $client->removeUser($user);
         if(0 === $user->getClients()->count() )
         {
             $this->userRepository->remove($user,true);
         }
+        return $user;
     }
 
 }
